refactor: Move stuff to gallia.utils
