+++ src/gallia/db/db_handler.py
@@ -7,9 +7,9 @@ from typing import Any, Optional
 
 import aiosqlite
 
+from gallia.penlog import Logger
 from gallia.uds.core import service
 from gallia.uds.core.utils import bytes_repr as bytes_repr_
-from gallia.penlog import Logger
 from gallia.utils import g_repr
