+++ src/gallia/udscan/scanner/scan_identifiers.py
@@ -11,12 +11,8 @@ from gallia.uds.core.service import NegativeResponse, UDSResponse
 from gallia.uds.core.utils import service_repr
 from gallia.uds.helpers import suggests_service_not_supported
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import (
-    ParseSkips,
-    auto_int,
-    check_and_set_session,
-)
-from gallia.utils import g_repr
+from gallia.udscan.utils import check_and_set_session
+from gallia.utils import ParseSkips, auto_int, g_repr
 
 
 class ScanIdentifiers(UDSScanner):
