+++ src/gallia/udscan/scanner/scan_reset.py
@@ -9,12 +9,8 @@ from gallia.uds.core.exception import IllegalResponse, UnexpectedNegativeRespons
 from gallia.uds.core.service import NegativeResponse, UDSResponse
 from gallia.uds.helpers import suggests_sub_function_not_supported
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import (
-    ParseSkips,
-    auto_int,
-    check_and_set_session,
-)
-from gallia.utils import g_repr
+from gallia.udscan.utils import check_and_set_session
+from gallia.utils import ParseSkips, auto_int, g_repr
 
 
 class ScanReset(UDSScanner):
