+++ src/gallia/udscan/scanner/scan_sa_dump_seeds.py
@@ -12,8 +12,8 @@ import aiofiles
 from gallia.uds.core.client import UDSRequestConfig
 from gallia.uds.core.service import NegativeResponse
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import auto_int, check_and_set_session
-from gallia.utils import g_repr
+from gallia.udscan.utils import check_and_set_session
+from gallia.utils import auto_int, g_repr
 
 
 class SaDumpSeeds(UDSScanner):
