+++ src/gallia/udscan/scanner/scan_services.py
@@ -10,12 +10,8 @@ from gallia.uds.core.exception import UDSException
 from gallia.uds.core.service import NegativeResponse, UDSResponse
 from gallia.uds.helpers import suggests_service_not_supported
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import (
-    ParseSkips,
-    auto_int,
-    check_and_set_session,
-)
-from gallia.utils import g_repr
+from gallia.udscan.utils import check_and_set_session
+from gallia.utils import ParseSkips, auto_int, g_repr
 
 
 class ScanServices(UDSScanner):
