+++ src/gallia/udscan/scanner/scan_sessions.py
@@ -11,8 +11,7 @@ from gallia.uds.core.service import (
     UDSResponse,
 )
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import auto_int
-from gallia.utils import g_repr
+from gallia.utils import auto_int, g_repr
 
 
 class IterateSessions(UDSScanner):
