+++ src/gallia/udscan/scanner/simple_ping.py
@@ -3,7 +3,7 @@ from argparse import Namespace
 
 from gallia.uds.core.service import NegativeResponse
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import auto_int
+from gallia.utils import auto_int
 
 
 class Ping(UDSScanner):
