+++ src/gallia/udscan/scanner/simple_read_error_log.py
@@ -3,8 +3,7 @@ from argparse import Namespace
 
 from gallia.uds.core.service import NegativeResponse
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import auto_int
-from gallia.utils import g_repr
+from gallia.utils import auto_int, g_repr
 
 
 class ReadErrorLog(UDSScanner):
