+++ src/gallia/udscan/scanner/simple_rtcl.py
@@ -6,8 +6,8 @@ from typing import Union
 
 from gallia.uds.core.service import NegativeResponse, RoutineControlResponse
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import auto_int, check_and_set_session
-from gallia.utils import g_repr
+from gallia.udscan.utils import check_and_set_session
+from gallia.utils import auto_int, g_repr
 
 
 class RTCL(UDSScanner):
