+++ src/gallia/udscan/scanner/simple_write_by_identifier.py
@@ -4,8 +4,7 @@ from argparse import Namespace
 
 from gallia.uds.core.service import NegativeResponse, UDSResponse
 from gallia.udscan.core import UDSScanner
-from gallia.udscan.utils import auto_int
-from gallia.utils import g_repr
+from gallia.utils import auto_int, g_repr
 
 
 class WriteByIdentifier(UDSScanner):
