+++ src/gallia/utils.py
@@ -1,11 +1,22 @@
 import ipaddress
 import re
+from argparse import Action, ArgumentError, ArgumentParser, Namespace
 from enum import Enum
-from typing import Optional, Any
-from urllib.parse import urlparse
+from pathlib import Path
+from sys import stdout
+from typing import Any, Callable, Optional, Sequence, Union
+from urllib.parse import ParseResult, urlencode, urlparse
 
-from gallia.uds.core.utils import bytes_repr, int_repr
+import aiofiles
+
+from gallia.db.db_handler import DBHandler
+from gallia.penlog import Logger
 from gallia.uds.core.service import NegativeResponse
+from gallia.uds.core.utils import bytes_repr, int_repr
+
+
+def auto_int(arg: str) -> int:
+    return int(arg, 0)
 
 
 def split_host_port(
@@ -82,3 +93,128 @@ def g_repr(x: Any) -> str:
         return str(x)
     else:
         return repr(x)
+
+
+def _unravel(listing: str) -> list[int]:
+    listing_delimiter = ","
+    range_delimiter = "-"
+    result = set()
+
+    for range_element in listing.split(listing_delimiter):
+        if range_delimiter in range_element:
+            first_tmp, last_tmp = range_element.split(range_delimiter)
+            first = auto_int(first_tmp)
+            last = auto_int(last_tmp)
+
+            for element in range(first, last + 1):
+                result.add(element)
+        else:
+            element = auto_int(range_element)
+            result.add(element)
+
+    return sorted(result)
+
+
+class ParseSkips(Action):
+    def __call__(
+        self,
+        parser: ArgumentParser,
+        namespace: Namespace,
+        values: Union[str, Sequence[Any], None],
+        option_string: str = None,
+    ) -> None:
+        skip_sids: dict[int, Optional[list[int]]] = {}
+
+        try:
+            if values is not None:
+                for session_skips in values:
+                    # Whole sessions can be skipped by only giving the session number without ids
+                    if ":" not in session_skips:
+                        session_ids = _unravel(session_skips)
+
+                        for session_id in session_ids:
+                            skip_sids[session_id] = None
+                    else:
+                        session_ids_tmp, identifier_ids_tmp = session_skips.split(":")
+                        session_ids = _unravel(session_ids_tmp)
+                        identifier_ids = _unravel(identifier_ids_tmp)
+
+                        for session_id in session_ids:
+                            if session_id not in skip_sids:
+                                skip_sids[session_id] = []
+
+                            session_skips = skip_sids[session_id]
+
+                            if session_skips is not None:
+                                session_skips += identifier_ids
+
+            setattr(namespace, self.dest, skip_sids)
+        except Exception as e:
+            raise ArgumentError(self, "The argument is malformed!") from e
+
+
+async def catch_and_log_exception(
+    logger: Logger,
+    func: Callable,
+    *args: Any,
+    **kwargs: Any,
+) -> None:
+    """Runs an async function. If an exception is raised,
+    it will be logged via logger.
+
+    :param logger: an instance of gallia.penlog.Logger
+    :param func: a async function object which will be awaited
+    :return: None
+    """
+    try:
+        return await func(*args, **kwargs)
+    except Exception as e:
+        logger.log_error(f"func {func.__name__} failed: {repr(e)}")
+
+
+class ANSIEscapes:
+    if stdout.isatty():
+        BOLD = "\033[1m"
+        ITALIC = "\033[3m"
+        UNDERSCORE = "\033[4m"
+        BLINK = "\033[5m"
+        CROSSED = "\033[9m"
+
+        BLACK = "\033[90m"
+        RED = "\033[91m"
+        GREEN = "\033[92m"
+        YELLOW = "\033[93m"
+        BLUE = "\033[94m"
+        MAGENTA = "\033[95m"
+        CYAN = "\033[96m"
+        WHITE = "\033[97m"
+
+        RESET = "\033[0m"
+    else:
+        BOLD = ITALIC = UNDERSCORE = BLINK = CROSSED = ""
+        BLACK = RED = GREEN = YELLOW = BLUE = MAGENTA = CYAN = WHITE = ""
+        RESET = ""
+
+
+async def write_target_list(
+    ecus_file: Path,
+    ecus: list[tuple[ParseResult, dict[str, Any]]],
+    db_handler: Optional[DBHandler] = None,
+) -> list[str]:
+    """Write a list of ECU connection strings (urls) into file
+
+    :param ecus_file: output file
+    :param ecus: list of ECUs with ECU specific url and params as dict
+    :params db_handler: if given, urls are also written to the database as discovery results
+    :return: None
+    """
+    urls: list[str] = list()
+    async with aiofiles.open(ecus_file, "w") as file:
+        for url, ecu in ecus:
+            url_result = f"{url.scheme}://{url.netloc}?{urlencode(ecu)}"
+            urls.append(url_result)
+            await file.write(f"{url_result}\n")
+
+            if db_handler is not None:
+                await db_handler.insert_discovery_result(url_result)
+    return urls
